# packages/contracts/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/modules": {
      "get": {
        "operationId": "AdminModulesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstalledModuleResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List installed modules, or only the ones with a newer published version",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminModulesController_install",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallModuleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallModuleResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Install a module by alias, optionally pinned to a version range",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/modules/{alias}": {
      "delete": {
        "operationId": "AdminModulesController_remove",
        "parameters": [
          {
            "name": "alias",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveModuleResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove an installed module that no started tournament references",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/modules/verify": {
      "post": {
        "operationId": "AdminModulesController_verify",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModuleVerifyResultResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Re-validate every installed module against the running core version",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "CopaLibre API",
    "description": "Self-hosted tournament management API.\n\nAuthentication is JWT Bearer, validated against the operator-selected\nidentity provider’s JWKS. Browsers use Authorization Code + PKCE and\nhold the access token in memory only — never localStorage. A token\npassed as a query parameter is ignored, because URLs leak into proxy\nlogs, browser history, metrics, traces and screenshots.\n\nEvery route declares one security plane: public-read,\nauthenticated-interaction, admin-control, or integration.\nauthenticated-interaction and admin-control share this same bearer\ntransport and differ only in authorization policy.",
    "version": "4.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Short-lived access token. Strict stateless mode is the default: no refresh credential is persisted, so a reload reauthenticates."
      }
    },
    "schemas": {
...
      "ProblemResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 403
          },
          "message": {
            "type": "string",
            "example": "subject may only act on their own records"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "InstalledModuleResponse": {
        "type": "object",
        "properties": {
          "moduleId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "enum": [
              "discipline",
              "tournament-profile"
            ]
          },
          "alias": {
            "type": "string",
            "example": "orbital-frisbee"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "sourceKind": {
            "type": "string",
            "enum": [
              "curated",
              "alternate"
            ]
          },
          "attributionAuthor": {
            "type": "string",
            "example": "SebaSOFT"
          }
        },
        "required": [
          "moduleId",
          "kind",
          "alias",
          "version",
          "sourceKind",
          "attributionAuthor"
        ]
      },
      "InstallModuleRequest": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "example": "orbital-frisbee"
          },
          "range": {
            "type": "string",
            "description": "Version range; defaults to the latest published",
            "example": "^1.0.0"
          },
          "source": {
            "type": "string",
            "description": "An explicitly allow-listed alternate source (COPALIBRE_MODULE_SOURCE_ALLOWLIST); omit to install from the curated repository",
            "example": "file:///var/lib/copalibre/modules-dev/orbital-frisbee"
          },
          "allowUnsatisfiedCapabilities": {
            "type": "boolean",
            "description": "Installs even when the declared required capabilities are not yet satisfied",
            "default": false
          }
        },
        "required": [
          "alias"
        ]
      },
      "InstallModuleResponse": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "discipline",
              "tournament-profile"
            ]
          },
          "alias": {
            "type": "string",
            "example": "orbital-frisbee"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "unsatisfiedRequiredCapabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "kind",
          "alias",
          "version",
          "unsatisfiedRequiredCapabilities"
        ]
      },
      "RemoveModuleResponse": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "example": "orbital-frisbee"
          },
          "removedCount": {
            "type": "number",
            "description": "Versions removed",
            "example": 1
          }
        },
        "required": [
          "alias",
          "removedCount"
        ]
      },
      "ModuleVerifyFailureResponse": {
        "type": "object",
        "properties": {
          "stage": {
            "type": "string",
            "example": "registry-reference"
          },
          "message": {
            "type": "string",
            "example": "Unregistered event code \"goal-x\""
          }
        },
        "required": [
          "stage",
          "message"
        ]
      },
      "ModuleVerifyResultResponse": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "example": "orbital-frisbee"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "ok": {
            "type": "boolean",
            "example": true
          },
          "failures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleVerifyFailureResponse"
            }
          }
        },
        "required": [
          "alias",
          "version",
          "ok",
          "failures"
        ]
      }
    }
  }
}